LayoutTests/fast/text-extraction/debug-text-extraction-offscreen-password-fields-expected.txt

-- texttree

root
    scrollable,scrollPosition=(0,0),contentSize=[…]
        section
            article
                'Article Title'
                'This is some visible content.'
        form,autocomplete='on',name='login'
            'Log in to your account'
            input,'text',placeholder='Username',name='username',required
            input,'password',placeholder='Password',name='password',required,secure
            input,'password',placeholder='Confirm Password',name='confirm-password',required,secure
            button,'Log in'
        'Enter access code'
        input,'password',placeholder='Access Code',name='access-code',secure
        form,autocomplete='on'
            'Create account'
            input,'text',placeholder='User',name='user',required
        input,'password',placeholder='Password',name='pass',required,secure

-- html

<body>
    <main>
        <section>
            <article>
                Article Title
                This is some visible content.
            </article>
        </section>
        <form autocomplete='on' name='login'>
            Log in to your account
            <input type='text' placeholder='Username' name='username' required>
            <input type='password' placeholder='Password' name='password' required>
            <input type='password' placeholder='Confirm Password' name='confirm-password' required>
            <button>Log in</button>
        </form>
        Enter access code
        <input type='password' placeholder='Access Code' name='access-code'>
        <form autocomplete='on'>
            Create account
            <input type='text' placeholder='User' name='user' required>
        </form>
        <input type='password' placeholder='Password' name='pass' required>
    </main>
</body>
